+++ tests/assets/url/renamed-filter.tmpl.js
@@ -10,10 +10,12 @@ export default function (_, { urlify, htmlUrlify }) {
         </head>
         <body>
             <a id="urlify" href="${urlify?.("/urlify/", true)}">Urlify</a>
-            ${htmlUrlify?.(
-    '<a id="htmlUrlify" href="/htmlUrlify/">htmlUrlify</a>',
-    true,
-  )}
+            ${
+    htmlUrlify?.(
+      '<a id="htmlUrlify" href="/htmlUrlify/">htmlUrlify</a>',
+      true,
+    )
+  }
         </body>
     </html>
   `;