+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Any BREAKING CHANGE between minor versions will be documented here in upper case
 - Use the wasm version of `lightningcss` due many bugs with the [N-API version in Deno](https://github.com/denoland/deno/issues/20072).
 - Urls with spaces [#481].
 - `on_demand` plugin: Fixed _preload.ts generation.
+- `picture` plugin: support object and array imagick data [#490].
 
 ## [1.18.5] - 2023-09-01
 ### Added