+++ CHANGELOG.md
@@ -28,7 +28,7 @@ Any BREAKING CHANGE between minor versions will be documented here in upper case
 
 ### Fixed
 - `minify_html` works offline.
-- Updated dependencies: `std`, `vento`, `terser`, `lightningcss`, `esbuild`, `sass`.
+- Updated dependencies: `std`, `vento`, `terser`, `lightningcss`, `esbuild`, `sass`, `postcss`, `autoprefixer`, `remark`.
 - Use the wasm version of `lightningcss` due many bugs with the [N-API version in Deno](https://github.com/denoland/deno/issues/20072).
 - Urls with spaces [#481].
 - `on_demand` plugin: Fixed _preload.ts generation.